requirements.txt
# Core application dependencies
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0

# Vector database and embeddings
qdrant-client>=1.7.0,<1.8.0
sentence-transformers>=2.2.0

# Knowledge graph database
neo4j>=5.0.0

# Data processing and utilities
pandas>=1.5.0
numpy>=1.24.0
tqdm>=4.65.0
typing-extensions>=4.8.0

# gRPC dependencies (specific versions for stability)
grpcio>=1.60.0,<1.61.0
grpcio-tools>=1.60.0,<1.61.0
protobuf>=4.21.0,<5.0.0

# PyTorch for sentence-transformers (CPU version)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Additional utilities that might be needed
requests>=2.28.0
urllib3>=1.26.0
certifi>=2022.12.7